Added the option to change screensavers from the settings page. Additionally, reorganized some type definitions.
