+++ client/src/router/index.ts
@@ -48,7 +48,7 @@ import Summary from "../views/terminal/weather/Summary.vue";
 import Whiteboard from "@/views/terminal/whiteboard/Whiteboard.vue";
 
 // Calculator Routes
-import Calculator from "@/views/screensaver/Bounce.vue";
+import Calculator from "@/views/terminal/calculator/Calculator.vue";
 
 // Atlas Routes
 import Atlas from "@/views/terminal/atlas/Atlas.vue";
