+++ client/src/types.ts
@@ -4,6 +4,27 @@
 
 import type {Nexus} from "@/views/terminal/nexus";
 
+export interface Preferences {
+    ui: {
+        screensaver: {
+            enabled: boolean
+            countdown: number
+            selection: string
+        }
+        background: {
+            image: string,
+            blur: boolean
+        }
+        theme: string
+        mode: string
+        blur: number
+        brightness: number
+        grid: boolean
+        watermark: boolean
+        night: boolean
+        outlines: boolean
+    }
+}
 
 export interface Controller {
     name: string
